+++ app/Http/Requests/Api/Auth/RegisterRequest.php
@@ -2,32 +2,10 @@
 
 namespace App\Http\Requests\Api\Auth;
 
-use App\Traits\ApiRespons;
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Http\Exceptions\HttpResponseException;
+use App\Http\Requests\ApiRequest;
 
-class RegisterRequest extends FormRequest
+class RegisterRequest extends ApiRequest
 {
-    use ApiRespons;
-
-    /**
-     * Indicates if the validator should stop on the first rule failure.
-     *
-     * @var bool
-     */
-    protected $stopOnFirstFailure = false;
-
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
     /**
      * Get the validation rules that apply to the request.
      *
@@ -36,59 +14,9 @@ class RegisterRequest extends FormRequest
     public function rules()
     {
         return [
-            'name' => ['required', 'string', 'max:255'],
-            'username' => ['required', 'string', 'max:255', 'unique:users,username'],
-            'password' => ['required', 'string', 'min:8', 'max:255', 'confirmed']
+            'name' => ['required','string','max:255'],
+            'username' => ['required','string','max:255','unique:users,username'],
+            'password' => ['required','string','min:8','max:255','confirmed']
         ];
-    }
-
-    /**
-     * Custom message for validation
-     *
-     * @return array
-     */
-    public function messages()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Get custom attributes for validator errors.
-     *
-     * @return array
-     */
-    public function attributes()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Prepare the data for validation.
-     *
-     * @return void
-     */
-    protected function prepareForValidation()
-    {
-        $this->merge([
-            // 
-        ]);
-    }
-
-    /**
-     * Custom error message for validation
-     *
-     * @return array
-     */
-    public function failedValidation(Validator $validator)
-    {
-        throw new HttpResponseException(
-            $this->createResponse('Server Error', route('api.register.store'), [
-                'data' => $validator->errors()
-            ], 400)
-        );
     }
 }
