+++ app/Http/Requests/Api/Publisher/StoreRequest.php
@@ -2,22 +2,10 @@
 
 namespace App\Http\Requests\Api\Publisher;
 
-use App\Traits\ApiRespons;
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Contracts\Validation\Validator;
-use Illuminate\Http\Exceptions\HttpResponseException;
+use App\Http\Requests\ApiRequest;
 
-class StoreRequest extends FormRequest
+class StoreRequest extends ApiRequest
 {
-    use ApiRespons;
-
-    /**
-     * Indicates if the validator should stop on the first rule failure.
-     *
-     * @var bool
-     */
-    protected $stopOnFirstFailure = false;
-
     /**
      * Determine if the user is authorized to make this request.
      *
@@ -38,55 +26,5 @@ class StoreRequest extends FormRequest
         return [
             'name' => ['required','string','max:255']
         ];
-    }
-
-    /**
-     * Custom message for validation
-     *
-     * @return array
-     */
-    public function messages()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Get custom attributes for validator errors.
-     *
-     * @return array
-     */
-    public function attributes()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Prepare the data for validation.
-     *
-     * @return void
-     */
-    protected function prepareForValidation()
-    {
-        $this->merge([
-            // 
-        ]);
-    }
-
-    /**
-     * Custom error message for validation
-     *
-     * @return array
-     */
-    public function failedValidation(Validator $validator)
-    {
-        throw new HttpResponseException(
-            $this->createResponse('Server Error', route('api.publisher.index'), [
-                'data' => $validator->errors()
-            ], 400)
-        );
     }
 }
