+++ app/Http/Requests/ApiRequest.php
@@ -1,0 +1,84 @@
+<?php
+
+namespace App\Http\Requests;
+
+use App\Traits\ApiRespons;
+use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Contracts\Validation\Validator;
+use Illuminate\Http\Exceptions\HttpResponseException;
+
+class ApiRequest extends FormRequest
+{
+    use ApiRespons;
+
+    /**
+     * Indicates if the validator should stop on the first rule failure.
+     *
+     * @var bool
+     */
+    protected $stopOnFirstFailure = false;
+
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return true;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array<string, mixed>
+     */
+    public function rules()
+    {
+        return [];
+    }
+
+    /**
+     * Custom message for validation
+     *
+     * @return array
+     */
+    public function messages()
+    {
+        return [];
+    }
+
+    /**
+     * Get custom attributes for validator errors.
+     *
+     * @return array
+     */
+    public function attributes()
+    {
+        return [];
+    }
+
+    /**
+     * Prepare the data for validation.
+     *
+     * @return void
+     */
+    protected function prepareForValidation()
+    {
+        $this->merge([]);
+    }
+
+    /**
+     * Custom error message for validation
+     *
+     * @return array
+     */
+    public function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(
+            $this->createResponse('Server Error', route('api.landing.index'), [
+                'data' => $validator->errors()
+            ], 400)
+        );
+    }
+}
