+++ app/Http/Requests/Web/Auth/RegisterRequest.php
@@ -2,28 +2,10 @@
 
 namespace App\Http\Requests\Web\Auth;
 
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Contracts\Validation\Validator;
+use App\Http\Requests\Request;
 
-class RegisterRequest extends FormRequest
+class RegisterRequest extends Request
 {
-    /**
-     * Indicates if the validator should stop on the first rule failure.
-     *
-     * @var bool
-     */
-    protected $stopOnFirstFailure = false;
-
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return true;
-    }
-
     /**
      * Get the validation rules that apply to the request.
      *
@@ -36,58 +18,5 @@ class RegisterRequest extends FormRequest
             'username' => ['required','max:255','string','unique:users,username'],
             'password' => ['required','min:8','max:255','string','confirmed']
         ];
-    }
-    
-
-    /**
-     * Custom message for validation
-     *
-     * @return array
-     */
-    public function messages()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Get custom attributes for validator errors.
-     *
-     * @return array
-     */
-    public function attributes()
-    {
-        return [
-            // 
-        ];
-    }
-
-    /**
-     * Prepare the data for validation.
-     *
-     * @return void
-     */
-    protected function prepareForValidation()
-    {
-        $this->merge([
-            // 
-        ]);
-    }
-    
-    /**
-     * Custom error response for validation.
-     *
-     * @return array
-     */
-    public function failedValidation(Validator $validator)
-    {
-        $toast = toastr();
-
-        foreach ($validator->messages()->all() as $message) {
-            $toast->error($message, 'Validation');
-        }
-
-        return $toast;
     }
 }
